error msg

+++ lib/domains/models/venue_info.dart
@@ -20,7 +20,7 @@ class VenueInfo with _$VenueInfo {
     @Default(None()) Option<String> microphones,
     @Default(None()) Option<String> lights,
     @Default(VenueType.other) VenueType type,
-    @Default(None()) Option<double> averageTicketPrice,
+    @Default(None()) Option<int> averageTicketPrice,
   }) = _VenueInfo;
 
   // fromJson
